LRCI-7362 Consolidate getBaseSlaveLabel in BaseTestClassGroup
