+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/test/clazz/group/BatchTestClassGroup.java
@@ -403,6 +403,7 @@ public abstract class BatchTestClassGroup extends BaseTestClassGroup {
 		return JenkinsMaster.getSlavesPerHostDefault();
 	}
 
+	@Override
 	public Integer getMinimumSlaveRAM() {
 		JobProperty jobProperty = getJobProperty(
 			"test.batch.minimum.slave.ram");
